add auth

+++ backend/Controllers/BudgetController.js
@@ -4,6 +4,7 @@ const addBudget = async (req, res) => {
   try {
     const { category, amount } = req.body;
     const budget = new BudgetModel({
+      user: req.user._id,
       category,
       amount,
     });
@@ -16,7 +17,7 @@ const addBudget = async (req, res) => {
 
 const getBudget = async (req, res) => {
   try {
-    const budget = await BudgetModel.find();
+    const budget = await BudgetModel.find({ user: req.user._id });
     res
       .status(200)
       .json({ message: "Budget fetched successfully", success: true, budget });
@@ -28,6 +29,19 @@ const getBudget = async (req, res) => {
 const deleteBudget = async (req, res) => {
   try {
     const { id } = req.params;
+
+    // Check if budget belongs to user
+    const budget = await BudgetModel.findById(id);
+
+    if (!budget) {
+      return res.status(404).json({ message: "Budget not found" });
+    }
+
+    // Make sure user owns budget
+    if (budget.user.toString() !== req.user._id.toString()) {
+      return res.status(401).json({ message: "Not authorized" });
+    }
+
     await BudgetModel.findByIdAndDelete(id);
     res
       .status(200)
